Add files via upload
Today the food industry has revolutionised due to the role played by food delivery apps. I have tried to inculcate my contribution in this revolution with the help of my food delivery app.
I have used Android Studio as IDE with the help of Flutter SDK and Pixel 2 as virtual device. The app will run on 98% of the device’s worldwide due to the use of Android Pie. 
This app enables us to order various types of our favourite food at the click of a finger.